requirements.txt
Django==1.8
django-crispy-forms==1.5.2
Pillow==3.0.0
42cc-pystyle
